player: custom fps now float
fps float only from 0.01 to 1, if more value then rounded to near int.

+++ examples/player/MainWindow.h
@@ -136,7 +136,7 @@ private slots:
 
     void onXunoBrowser();
     void onClickXunoBrowser();
-    void customfpsChanged(int n);
+    void customfpsChanged(double n);
 
 protected:
     virtual void closeEvent(QCloseEvent *e);
@@ -164,7 +164,7 @@ private:
     int mCursorTimer;
     int mShowControl; //0: can hide, 1: show and playing, 2: always show(not playing)
     int mRepeateMax;
-    int mCustomFPS;
+    double mCustomFPS;
     QVBoxLayout *mpPlayerLayout;
 
     QWidget *mpControl;
